+++ client/components/devtools.js
@@ -2,222 +2,191 @@ import React, { Component } from "react";
 import { Meteor } from "meteor/meteor";
 import { composeWithTracker, registerComponent } from "@reactioncommerce/reaction-components";
 import { Button, SettingsCard } from "@reactioncommerce/reaction-ui";
+import { Components } from "@reactioncommerce/reaction-components";
 
 
 class DevTools extends Component {
-  handleResetDataClick = () => {
-    Meteor.call("devtools/resetData", (error) => {
+
+  state = {
+    mongoHost: "mongodb://localhost:3001",
+    dbName: "meteor",
+    products: 100,
+    orders: 100,
+    users: 100,
+    tags: 3,
+    variations: [1, 2, 4, 5],
+    IPS: 3,
+    discounts: 10,
+    attributes: 10,
+    duration: 1,
+    productBatchSize: 3001,
+    imageBatchSize: 3000,
+    optionsImageCached: true,
+    primaryImageCached: true
+  };
+
+  populateSmallDataset = () => {
+    this.setState({
+      products: 100,
+      orders: 100,
+      tags: 3,
+      IPS: 3,
+    })
+  }
+
+  populateMediumDataset = () => {
+    this.setState({
+      products: 1000,
+      orders: 10000,
+      tags: 10,
+      imageBatchSize: 5000,
+      IPS: 5
+    })
+  }
+
+  populateLargeDataset = () => {
+    this.setState({
+      products: 50000,
+      orders: 50000,
+      tags: 100,
+      imageBatchSize: 3000,
+      IPS: 7
+    })
+  }
+
+  populateSampleDataset = () => {
+    Meteor.call("devtools/sampleData", this.state, (error) => {
       if (error) {
-        Alerts.toast(`Error resetting sample data ${error.reason}`, "error");
+        Alerts.toast(`Error loading sample data ${error.msg}`, "error");
+      } else {
+        Alerts.toast("Sample data load successful", "success");
+      }
+    });
+  }
+
+  handleResetData = () => {
+    Meteor.call("devtools/resetData", this.state, (error) => {
+      if (error) {
+        Alerts.toast(`Error resetting sample data ${error.msg}`, "error");
       } else {
         Alerts.toast("Reset successful", "success");
       }
     });
   }
 
-  handleSwagShopImagesClick = () => {
-    Meteor.call("devtools/loaddata/images", "swagshop", (error) => {
+  handleLoadData = () => {
+    Meteor.call("devtools/loadData", this.state, (error) => {
       if (error) {
-        Alerts.toast(`Error loading images ${error.reason}`, "error");
+        Alerts.toast(`Error loading sample data ${error.msg}`, "error");
       } else {
-        Alerts.toast("Images loaded successfully", "success");
+        Alerts.toast("Data loading successful", "success");
       }
     });
   }
 
-  handleImagesClick = () => {
-    Meteor.call("devtools/loaddata/images", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading images ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Images loaded successfully", "success");
-      }
-    });
-  }
-
-  handleImagesFromWebClick = () => {
-    Meteor.call("devtools/loaddata/images", "web", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading images ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Images loaded successfully", "success");
-      }
-    });
-  }
-
-  handleSeedDataClick = () => {
-    Meteor.call("devtools/loaddata/small/products", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading sample data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Sample data loaded", "success");
-      }
-    });
-  }
-
-  handleSmallOrdersClick = () => {
-    Meteor.call("devtools/loaddata/small/orders", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading order data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Orders loaded successfully", "success");
-      }
-    });
-  }
-
-  handleMediumDataClick = () => {
-    Meteor.call("devtools/loaddata/medium/products", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading medium sample data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Load medium dataset successful", "success");
-      }
-    });
-  }
-
-  handleMediumOrdersClick = () => {
-    Meteor.call("devtools/loaddata/medium/orders", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading medium sample data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Loaded 10000 orders successful", "success");
-      }
-    });
-  }
-
-  handleLargeDataClick = () => {
-    Meteor.call("devtools/loaddata/large/products", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading large sample data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Load large dataset successful", "success");
-      }
-    });
-  }
-
-  handleLargeOrdersClick = () => {
-    Meteor.call("devtools/loaddata/large/orders", (error) => {
-      if (error) {
-        Alerts.toast(`Error loading large order data ${error.reason}`, "error");
-      } else {
-        Alerts.toast("Loading orders successful", "success");
-      }
-    });
+  handleChange = (event, value, name) => {
+    this.setState({
+      [name]: value
+    })
   }
 
   render() {
     return (
-      <div>
+      <div className="devtools">
         <SettingsCard
           title={"Common"}
           expanded={true}
           showSwitch={false}
         >
+        <div id="datasets">
+          <Button
+            bezelStyle={"solid"}
+            label={"Small Dataset"}
+            onClick={this.populateSmallDataset}
+            className={"small"}
+          />
+          <Button
+            bezelStyle={"solid"}
+            label={"Medium Dataset"}
+            onClick={this.populateMediumDataset}
+            className={"medium"}
+          />
+          <Button
+            bezelStyle={"solid"}
+            label={"Large Dataset"}
+            onClick={this.populateLargeDataset}
+            className={"large"}
+          />
+        </div>
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Database name"}
+          placeholder={"meteor"}
+          value={this.state.dbName}
+          name={"dbName"}
+        />
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Mongo host"}
+          placeholder={"mongodb://localhost:3001"}
+          value={this.state.mongoHost}
+          name={"mongoHost"}
+        />
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Top level products"}
+          placeholder={1000}
+          type={"number"}
+          value={this.state.products}
+          name={"products"}
+        />
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Orders"}
+          placeholder={1000}
+          type={"number"}
+          value={this.state.orders}
+          name={"orders"}
+        />
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Categories"}
+          placeholder={10}
+          type={"number"}
+          value={this.state.tags}
+          name={"tags"}
+        />
+        <Components.TextField
+          onChange={this.handleChange}
+          label={"Images per SKU"}
+          placeholder={3}
+          type={"number"}
+          value={this.state.IPS}
+          name={"IPS"}
+        />
+        <div id="actionButtons">
+          <Button
+            bezelStyle={"solid"}
+            primary={true}
+            label={"Load Data"}
+            onClick={this.handleLoadData}
+            className={"loadData"}
+          />
           <Button
             bezelStyle={"solid"}
             primary={true}
             label={"Reset Data"}
-            onClick={this.handleResetDataClick}
+            onClick={this.handleResetData}
+            className={"resetData"}
           />
-        </SettingsCard>
-
-        <SettingsCard
-          title={"Small shop data (10 products, 100 orders"}
-          expanded={true}
-          showSwitch={false}
-        >
           <Button
             bezelStyle={"solid"}
+            label={"Load Sample Dataset"}
             primary={true}
-            label={"Load Products and Tags"}
-            onClick={this.handleSeedDataClick}
+            onClick={this.populateSampleDataset}
+            className={"loadSampleData"}
           />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Orders"}
-            onClick={this.handleSmallOrdersClick}
-          />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Images"}
-            onClick={this.handleSwagShopImagesClick}
-          />
-
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Puppy Images"}
-            onClick={this.handleImagesFromWebClick}
-          />
-        </SettingsCard>
-
-
-        <SettingsCard
-          title={"Medium Dataset (1000 products, 10000 orders)"}
-          expanded={true}
-          showSwitch={false}
-        >
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Produts and Tags"}
-            onClick={this.handleMediumDataClick}
-          />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Orders"}
-            onClick={this.handleMediumOrdersClick}
-          />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Images"}
-            onClick={this.handleImagesClick}
-          />
-          <div>Loading images may take some time</div>
-        </SettingsCard>
-
-        <SettingsCard
-          title={"Large Dataset (50,000 products, 50,000 orders)"}
-          expanded={true}
-          showSwitch={false}
-        >
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Produts and Tags"}
-            onClick={this.handleLargeDataClick}
-          />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Orders"}
-            onClick={this.handleLargeOrdersClick}
-          />
-          <br />
-          <br />
-          <Button
-            bezelStyle={"solid"}
-            primary={true}
-            label={"Load Images"}
-            onClick={this.handleImagesClick}
-          />
-          <div>Loading images may take some time</div>
+        </div>
         </SettingsCard>
       </div>
     );
