+++ client/devtools.less
@@ -1,0 +1,27 @@
+.devtools {
+    #datasets {
+        text-align: center;
+        margin-bottom: 12px;
+        .small {
+            float: left;
+        }
+        .medium {
+            display: inline-block;
+        }
+        .large {
+            float: right;
+        }
+    }
+    #actionButtons {
+        text-align: center;
+        .loadData {
+            float: left;
+        }
+        .resetData {
+            display: inline-block;
+        }
+        .loadSampleData {
+            float: right;
+        }
+    }
+}
