+++ client/index.js
@@ -1,1 +1,2 @@
+import "./devtools.less";
 export { default as DevTools } from "./components/devtools";
