--- data-loader/.babelrc
@@ -1,4 +1,0 @@
-{
-    "presets": ["es2015", "stage-2"],
-    "plugins": []
-  }
