--- data-loader/config/scripts/reaction_mid
@@ -1,3 +1,0 @@
-#!/bin/bash
-MONGO_URL=mongodb://172.31.18.209:27017/mid ROOT_URL=http://localhost.com PORT=5000 node --max_old_space_size=32768 /home/ec2-user/build/bundle/main.js > mid.log 2>mid_errors.log
-
