--- data-loader/config/scripts/reaction_ret
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-MONGO_URL=mongodb://172.31.18.209:27017/ret ROOT_URL=http://localhost.com PORT=4000 node --max_old_space_size=32768 /home/ec2-user/build/bundle/main.js > ret.log 2> ret_errors.log
-
