--- data-loader/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "reaction-developer-tools",
-  "version": "0.0.1",
-  "description": "Reaction inline developer tools",
-  "main": "register.js",
-  "author": "Reaction Commerce <hello@reactioncommerce.com>",
-  "license": "GPL-3.0",
-  "homepage": "https://github.com/reactioncommerce",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/reactioncommerce/reaction-devtools.git"
-  },
-  "dependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "buffer-stream-reader": "^0.1.1",
-    "faker": "^4.1.0",
-    "mongodb": "^3.0.6",
-    "nodemon": "1.17.3",
-    "node-random": "^1.0.1",
-    "performance-now": "^2.1.0",
-    "random-id": "0.0.2",
-    "sharp": "^0.20.1",
-    "slugify": "^1.2.9",
-    "uuid": "^3.2.1",
-    "webworker-threads": "^0.7.13"
-  },
-  "scripts": {
-    "lint": "eslint .",
-    "start": "nodemon data/generator.js --exec babel-node"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "eslint": "^4.17.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.6.1"
-  }
-}
