+++ sample-data/Products.json
@@ -1039,4 +1039,3 @@
 	},
 	"originCountry": "US"
 }]
-
