+++ sample-data/Shops.json
@@ -3495,4 +3495,3 @@
 		}
 	]
 }]
-
