+++ sample-data/Tags.json
@@ -18,4 +18,3 @@
 	"groups": [],
 	"shopId": "J8Bhq3uTtdgwZx3rz"
 }]
-
